split all attributes stuff in a own class
