+++ classes/class-algolia-woo-indexer.php
@@ -12,6 +12,7 @@ namespace Algowoo;
 use \Algowoo\Algolia_Check_Requirements as Algolia_Check_Requirements;
 use \Algowoo\Algolia_Verify_Nonces as Algolia_Verify_Nonces;
 use \Algowoo\Algolia_Send_Products as Algolia_Send_Products;
+use \Algowoo\Algolia_Attributes as Algolia_Attributes;
 
 /**
  * Abort if this file is called directly
@@ -156,24 +157,10 @@ if (!class_exists('Algolia_Woo_Indexer')) {
                 );
 
                 /**
-                 * Add sections and fields for the attributes
+                 * add settings for Attributes
                  */
-                add_settings_section(
-                    'algolia_woo_indexer_attributes',
-                    esc_html__('Attributes indexing settings', 'algolia-woo-indexer'),
-                    array($algowooindexer, 'algolia_woo_indexer_attributes_section_text'),
-                    'algolia_woo_indexer'
-                );
+                Algolia_Attributes::setup_attributes_settings();
 
-                foreach (ATTRIBUTES_SETTINGS as $key => $description) {
-                    add_settings_field(
-                        'algolia_woo_indexer_attributes_' . $key,
-                        esc_html__($description, 'algolia-woo-indexer'),
-                        array($algowooindexer, 'algolia_woo_indexer_attributes_' . $key . '_output'),
-                        'algolia_woo_indexer',
-                        'algolia_woo_indexer_attributes'
-                    );
-                }
             }
         }
 
@@ -270,148 +257,6 @@ if (!class_exists('Algolia_Woo_Indexer')) {
         }
 
         /**
-         * Output for attributes if functionality is enabled
-         *
-         * @return void
-         */
-        public static function algolia_woo_indexer_attributes_enabled_output()
-        {
-            $value = get_option(ALGOWOO_DB_OPTION . ATTRIBUTES_ENABLED);
-            $isChecked = (!empty($value)) ? 1 : 0;
-        ?>
-            <input id="algolia_woo_indexer_attributes_enabled" name="algolia_woo_indexer_attributes_enabled[checked]" type="checkbox" <?php checked(1, $isChecked); ?> />
-            <?php
-        }
-
-        /**
-         * Output for attributes how to handle visibility setting
-         *
-         * @return void
-         */
-        public static function algolia_woo_indexer_attributes_visibility_output()
-        {
-            $value = get_option(ALGOWOO_DB_OPTION . ATTRIBUTES_VISIBILITY);
-            foreach (ATTRIBUTES_VISIBILITY_STATES as $state) {
-                $id = 'algolia_woo_indexer_attributes_visibility_' . $state;
-            ?>
-                <p><input id="<?php echo $id; ?>" name="algolia_woo_indexer_attributes_visibility[value]" type="radio" value="<?php echo $state; ?>" <?php checked($state, $value); ?> /><label for="<?php echo $id; ?>"><?php echo esc_html__($state, 'algolia-woo-indexer'); ?></label></p>
-            <?php
-            }
-        }
-
-        /**
-         * Output for attributes how to handle variant setting
-         *
-         * @return void
-         */
-        public static function algolia_woo_indexer_attributes_variation_output()
-        {
-            $value = get_option(ALGOWOO_DB_OPTION . ATTRIBUTES_VARIATION);
-            foreach (ATTRIBUTES_VARIATION_STATES as $state) {
-                $id = 'algolia_woo_indexer_attributes_variation_' . $state;
-            ?>
-                <p><input id="<?php echo $id; ?>" name="algolia_woo_indexer_attributes_variation[value]" type="radio" value="<?php echo $state; ?>" <?php checked($state, $value); ?> /><label for="<?php echo $id; ?>"><?php echo esc_html__($state, 'algolia-woo-indexer'); ?></label></p>
-            <?php
-            }
-        }
-
-        /**
-         * Output for attributes list which attributes are whitelisted
-         *
-         * @return void
-         */
-        public static function algolia_woo_indexer_attributes_list_output()
-        {
-            $value = get_option(ALGOWOO_DB_OPTION . ATTRIBUTES_LIST);
-            $selectedIds = explode(",", $value);
-            $name = "algolia_woo_indexer_attributes_list[list]";
-            $description = __('Here you can whitelist all the attributes. Use the <b>shift</b> or <b>control</b> buttons to select multiple attributes.', 'algolia-woo-indexer');
-            self::generic_attributes_select_output($name, $selectedIds, $description);
-        }
-
-        /**
-         * Output for attributes list which are using a numeric interpolation
-         *
-         * @return void
-         */
-        public static function algolia_woo_indexer_attributes_interp_output()
-        {
-            $value = get_option(ALGOWOO_DB_OPTION . ATTRIBUTES_INTERP);
-            $selectedIds = explode(",", $value);
-            $name = "algolia_woo_indexer_attributes_interp[list]";
-            $description = __('If you have some attributes based on number which shall be interpd between the lowest to the highest number, you can select it here. A common usecase for this is if you want to have a <b>range slider</b> in aloglia which works for a certain range. Example: a plant grows between 20 and 25cm tall. for this you enter 20 and 25 as attribute values to your product and it will automatically extend the data to [20,21,22,23,24,25]', 'algolia-woo-indexer');
-            self::generic_attributes_select_output($name, $selectedIds, $description);
-        }
-
-        /**
-         * Output for attributes list which are using a numeric interpolation
-         *
-         * @return void
-         */
-        public static function algolia_woo_indexer_attributes_tax_fields_output()
-        {
-            $selected_raw = get_option(ALGOWOO_DB_OPTION . ATTRIBUTES_TAX_FIELDS);
-            $selected_entries = explode(",", $selected_raw);
-            $name = "algolia_woo_indexer_attributes_tax_fields[list]";
-            $description = __('Select which taxonomy fields for each attribute shall be indexed', 'algolia-woo-indexer');
-            $values = ALLOWED_TAXONOMIES;
-            ?>
-            <p><?php echo $description; ?></p>
-            <select multiple="multiple" name="<?php echo $name; ?>[]" size="<?php echo count($values); ?>">
-                <?php
-                foreach ($values as $tax) {
-                    $selected = in_array($tax, $selected_entries) ? ' selected="selected" ' : '';
-                ?>
-                    <option value="<?php echo $tax; ?>" <?php echo $selected; ?>>
-                        <?php echo __($tax, 'algolia-woo-indexer'); ?>
-                    </option>
-                <?php
-                }
-                ?>
-            </select>
-        <?php
-        }
-
-        /**
-         * Generic Output for attributes list where attributes are whitelisted using Woocommerce attributes taxonomies
-         * @param string $name id and name for select
-         * @param array $selected_entries will be preselected if matching with WC taxonomies
-         * @param string $description will be displayed on top
-         * 
-         */
-        public static function generic_attributes_select_output($name, $selected_entries, $description)
-        {
-
-
-            $values = wc_get_attribute_taxonomies();
-            if (!$values) {
-                echo esc_html__('You don\'t have any attributes defined yet. Go to WooCommerce and add some to use this feature.', 'algolia-woo-indexer');
-                return;
-            }
-
-        ?>
-            <p><?php echo $description; ?></p>
-            <select multiple="multiple" name="<?php echo $name; ?>[]" size="<?php echo count($values); ?>">
-                <?php
-                foreach ($values as $tax) {
-
-                    $id = $tax->attribute_id;
-                    $label = $tax->attribute_label;
-                    $name = $tax->attribute_name;
-                    $selected = in_array($id, $selected_entries) ? ' selected="selected" ' : '';
-                ?>
-                    <option value="<?php echo $id; ?>" <?php echo $selected; ?>>
-                        <?php echo $label . ' (' . $name . ')'; ?>
-                    </option>
-                <?php
-                }
-                ?>
-            </select>
-        <?php
-
-        }
-
-        /**
          * Section text for plugin settings section text
          *
          * @return void
@@ -430,17 +275,7 @@ if (!class_exists('Algolia_Woo_Indexer')) {
         {
             echo esc_html__('Choose which basic fields shall be indexed.', 'algolia-woo-indexer');
         }
-
-        /**
-         * Section text for attributes settings section text
-         *
-         * @return void
-         */
-        public static function algolia_woo_indexer_attributes_section_text()
-        {
-            echo esc_html__('Control if and how the attributes shall be indexed.', 'algolia-woo-indexer');
-        }
-
+        
         /**
          * Check if we are going to send products by verifying send products nonce
          *
@@ -559,7 +394,7 @@ if (!class_exists('Algolia_Woo_Indexer')) {
             }
 
             /**
-             * Filter the application id, api key, index name and verify that the input is an array
+             * Filter the application id, api key, index name, custom_fields and verify that the input is an array
              *
              * @see https://www.php.net/manual/en/function.filter-input.php
              */
@@ -568,12 +403,7 @@ if (!class_exists('Algolia_Woo_Indexer')) {
             $post_index_name                 = filter_input(INPUT_POST, 'algolia_woo_indexer_index_name', FILTER_DEFAULT, FILTER_REQUIRE_ARRAY);
             $auto_send                       = filter_input(INPUT_POST, 'algolia_woo_indexer_automatically_send_new_products', FILTER_DEFAULT, FILTER_REQUIRE_ARRAY);
             $custom_fields                   = filter_input(INPUT_POST, 'algolia_woo_indexer_custom_fields', FILTER_DEFAULT);
-            $attributes_enabled              = filter_input(INPUT_POST, 'algolia_woo_indexer_attributes_enabled', FILTER_DEFAULT, FILTER_REQUIRE_ARRAY);
-            $attributes_visibility           = filter_input(INPUT_POST, 'algolia_woo_indexer_attributes_visibility', FILTER_DEFAULT, FILTER_REQUIRE_ARRAY);
-            $attributes_variation            = filter_input(INPUT_POST, 'algolia_woo_indexer_attributes_variation', FILTER_DEFAULT, FILTER_REQUIRE_ARRAY);
-            $attributes_list                 = filter_input(INPUT_POST, 'algolia_woo_indexer_attributes_list', FILTER_DEFAULT, FILTER_REQUIRE_ARRAY);
-            $attributes_interp               = filter_input(INPUT_POST, 'algolia_woo_indexer_attributes_interp', FILTER_DEFAULT, FILTER_REQUIRE_ARRAY);
-            $attributes_tax_fields               = filter_input(INPUT_POST, 'algolia_woo_indexer_attributes_tax_fields', FILTER_DEFAULT, FILTER_REQUIRE_ARRAY);
+
             /**
              * Properly sanitize text fields before updating data
              *
@@ -584,43 +414,11 @@ if (!class_exists('Algolia_Woo_Indexer')) {
             $sanitized['api_key']                   = sanitize_text_field($post_api_key['key']);
             $sanitized['index_name']                = sanitize_text_field($post_index_name['name']);
             $sanitized['custom_fields']             = sanitize_textarea_field($custom_fields);
-            $sanitized['attributes_visibility']     = sanitize_text_field($attributes_visibility['value']);
-            $sanitized['attributes_variation']      = sanitize_text_field($attributes_variation['value']);
-
-            /**
-             * sanitize select list of id's by getting integers and them implode seperated with comma
-             */
-
-            $attributes_list_integers = [];
-            foreach ($attributes_list['list'] as $id) {
-                $sanitizedId = sanitize_text_field($id);
-                array_push($attributes_list_integers, (int) $sanitizedId);
-            }
-            $sanitized['attributes_list'] = implode(',', $attributes_list_integers);
-
-            $attributes_interp_int = [];
-            foreach ($attributes_interp['list'] as $id) {
-                $sanitizedId = sanitize_text_field($id);
-                array_push($attributes_interp_int, (int) $sanitizedId);
-            }
-            $sanitized['attributes_interp'] = implode(',', $attributes_interp_int);
-
-            /**
-             * only allow values from the ALLOWED_TAXONOMIES to be saved
-             */
-            $sanitized['attributes_tax_fields'] = [];
-            foreach ($attributes_tax_fields['list'] as $name) {
-                if (in_array($name, ALLOWED_TAXONOMIES)) {
-                    array_push($sanitized['attributes_tax_fields'], $name);
-                }
-            }
-            $sanitized['attributes_tax_fields'] = implode(',', $sanitized['attributes_tax_fields']);
 
             /**
              * Sanitizing by setting the value to either 1 or 0
              */
             $sanitized['product'] = (!empty($auto_send)) ? 1 : 0;
-            $sanitized['attributes_enabled'] = (!empty($attributes_enabled)) ? 1 : 0;
 
             /**
              * Filter the data fields checkboxes and verify that the input is an array and assign it to an associative array
@@ -667,26 +465,6 @@ if (!class_exists('Algolia_Woo_Indexer')) {
                     $sanitized['product']
                 );
             }
-            if (isset($sanitized['custom_fields'])) {
-                $custom_fields_array = preg_split("/[\r\n,]+/", $sanitized['custom_fields'], -1, PREG_SPLIT_NO_EMPTY);
-                $custom_fields_string = implode(",", $custom_fields_array);
-                update_option(
-                    ALGOWOO_DB_OPTION . CUSTOM_FIELDS,
-                    $custom_fields_string
-                );
-            }
-
-            foreach (array_keys(ATTRIBUTES_SETTINGS) as $key) {
-                $value = $sanitized['attributes_' . $key];
-                if (isset($value)) {
-                    $extension = constant('ATTRIBUTES_' . strtoupper($key));
-                    var_dump($extension);
-                    update_option(
-                        ALGOWOO_DB_OPTION . $extension,
-                        $value
-                    );
-                }
-            }
 
             if (isset($sanitized['fields']) && (!empty($sanitized['fields']))) {
                 foreach ($sanitized['fields'] as $key => $value) {
@@ -696,6 +474,21 @@ if (!class_exists('Algolia_Woo_Indexer')) {
                     );
                 }
             }
+
+            if (isset($sanitized['custom_fields'])) {
+                $custom_fields_array = preg_split("/[\r\n,]+/", $sanitized['custom_fields'], -1, PREG_SPLIT_NO_EMPTY);
+                $custom_fields_string = implode(",", $custom_fields_array);
+                update_option(
+                    ALGOWOO_DB_OPTION . CUSTOM_FIELDS,
+                    $custom_fields_string
+                );
+            }
+
+            /**
+             * Update Attributes as well 
+             */
+            Algolia_Attributes::update_attribute_options();
+
         }
 
         /**
@@ -801,12 +594,6 @@ if (!class_exists('Algolia_Woo_Indexer')) {
             $algolia_api_key                 = get_option(ALGOWOO_DB_OPTION . ALGOLIA_API_KEY);
             $algolia_index_name              = get_option(ALGOWOO_DB_OPTION . INDEX_NAME);
             $custom_fields                   = get_option(ALGOWOO_DB_OPTION . CUSTOM_FIELDS);
-            $attributes_enabled              = get_option(ALGOWOO_DB_OPTION . ATTRIBUTES_ENABLED);
-            $attributes_visibility           = get_option(ALGOWOO_DB_OPTION . ATTRIBUTES_VISIBILITY);
-            $attributes_variation            = get_option(ALGOWOO_DB_OPTION . ATTRIBUTES_VARIATION);
-            $attributes_list                 = get_option(ALGOWOO_DB_OPTION . ATTRIBUTES_LIST);
-            $attributes_interp               = get_option(ALGOWOO_DB_OPTION . ATTRIBUTES_INTERP);
-            $attributes_tax_fields           = get_option(ALGOWOO_DB_OPTION . ATTRIBUTES_TAX_FIELDS);
 
             if (empty($auto_send)) {
                 add_option(
@@ -842,44 +629,6 @@ if (!class_exists('Algolia_Woo_Indexer')) {
                 );
             }
 
-            if (empty($attributes_enabled)) {
-                add_option(
-                    ALGOWOO_DB_OPTION . ATTRIBUTES_ENABLED,
-                    0
-                );
-            }
-
-            if (empty($attributes_visibility)) {
-                add_option(
-                    ALGOWOO_DB_OPTION . ATTRIBUTES_VISIBILITY,
-                    ATTRIBUTES_VISIBILITY_STATES[1]
-                );
-            }
-            if (empty($attributes_variation)) {
-                add_option(
-                    ALGOWOO_DB_OPTION . ATTRIBUTES_VARIATION,
-                    ATTRIBUTES_VARIATION_STATES[0]
-                );
-            }
-            if (empty($attributes_list)) {
-                add_option(
-                    ALGOWOO_DB_OPTION . ATTRIBUTES_LIST,
-                    ''
-                );
-            }
-            if (empty($attributes_interp)) {
-                add_option(
-                    ALGOWOO_DB_OPTION . ATTRIBUTES_INTERP,
-                    ''
-                );
-            }
-            if (empty($attributes_tax_fields)) {
-                add_option(
-                    ALGOWOO_DB_OPTION . ATTRIBUTES_TAX_FIELDS,
-                    'name,slug'
-                );
-            }
-
             /**
              * Set default values for index fields if not already set
              */
@@ -893,7 +642,10 @@ if (!class_exists('Algolia_Woo_Indexer')) {
                 }
             }
 
-
+            /**
+             * activate attributes
+             */
+            Algolia_Attributes::activate_attributes();
 
             set_transient(self::PLUGIN_TRANSIENT, true);
         }
